myBlogSite/requirements.txt
django_ckeditor==6.1.0
Django==3.2.4
